rust: Remove potential segmentation fault in time
Previous we were using the crate time.  The version we used contained
an issue which could cause a segmentation fault, ref.
https://github.com/advisories/GHSA-wcg3-cvx6-7396

Remove our use of time and remove our use of crate rust-crypto which
hasn't been updated in a very long time and also used an old version
of time library.  Replaced with md5 library which hasn't been updated
for a while, but has no dependencies which is rare in the rust
world.

Signed-off-by: Tony Asleson <tasleson@redhat.com>